changed page organization
