+++ client/src/pages/Login.tsx
@@ -1,10 +1,10 @@
 import { useContext, useEffect } from 'react';
-import { AuthContext } from '../../context/auth/authProvider';
+import { AuthContext } from '../context/auth/authProvider';
 import { useNavigate } from 'react-router-dom';
-import styles from '../../styles/pages/Register.module.scss';
-import RegisterForm from '../../components/RegisterForm';
+import LoginForm from '../components/LoginForm';
+import styles from '../styles/pages/Login.module.scss';
 
-const Register = () => {
+const Login = () => {
 	const { isAuthenticated } = useContext(AuthContext);
 	let navigate = useNavigate();
 
@@ -15,10 +15,10 @@ const Register = () => {
 	}, [isAuthenticated, navigate]);
 
 	return (
-		<div className={styles.register}>
-			<RegisterForm />
+		<div className={styles.login}>
+			<LoginForm />
 		</div>
 	);
 };
 
-export default Register;
+export default Login;
