+++ client/src/pages/Register.tsx
@@ -1,10 +1,10 @@
 import { useContext, useEffect } from 'react';
-import { AuthContext } from '../../context/auth/authProvider';
+import { AuthContext } from '../context/auth/authProvider';
 import { useNavigate } from 'react-router-dom';
-import LoginForm from '../../components/LoginForm';
-import styles from '../../styles/pages/Login.module.scss';
+import styles from '../styles/pages/Register.module.scss';
+import RegisterForm from '../components/RegisterForm';
 
-const Login = () => {
+const Register = () => {
 	const { isAuthenticated } = useContext(AuthContext);
 	let navigate = useNavigate();
 
@@ -15,10 +15,10 @@ const Login = () => {
 	}, [isAuthenticated, navigate]);
 
 	return (
-		<div className={styles.login}>
-			<LoginForm />
+		<div className={styles.register}>
+			<RegisterForm />
 		</div>
 	);
 };
 
-export default Login;
+export default Register;
