+++ client/src/pages/Scores.tsx
